+++ recipes/nfs_server.rb
@@ -1,9 +1,10 @@
 # encoding: utf-8
-#
 # Cookbook Name:: chef-splunk
-# Recipe:: default
-#
+# Recipe:: nfs_server
 # Author: Bernardo Macias <bmacias84@gmail.com>
+#
+#
+# All rights reserved - Do Not Redistribute
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,16 +17,24 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
 
-if node[:splunk][:disabled]
-  include_recipe 'chef-splunk::disabled'
-  Chef::Log.debug('Splunk is disabled on this node.')
-  return
+include_recipe 'nfs::server'
+
+nfs_mnt = node[:splunk][:searchpool][:pool_mnt]
+Chef::Log.debug(nfs_mnt)
+directory 'nfs_mnt' do
+  owner  node[:splunk][:dir_perms][:owner]
+  group  node[:splunk][:dir_perms][:group]
+  mode   node[:splunk][:dir_perms][:mode]
+  action :create
 end
 
-if node[:splunk][:is_server]
-  include_recipe "chef-splunk::#{node[:splunk][:type]}"
-else
-  include_recipe 'chef-splunk::client'
+export_name = %r{\/[^\/]+$}.match(nfs_mnt).to_s
+
+nfs_export export_name do
+  directory nfs_mnt
+  network   node[:splunk][:searchpool][:network]
+  writeable true
+  sync      true
+  options   %w{no_root_squash no_subtree_check nohide no_wdelay}
 end