+++ recipes/search.rb
@@ -1,9 +1,8 @@
 # encoding: utf-8
-#
 # Cookbook Name:: chef-splunk
-# Recipe:: default
-#
+# Recipe:: search
 # Author: Bernardo Macias <bmacias84@gmail.com>
+#
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,16 +15,22 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
 
-if node[:splunk][:disabled]
-  include_recipe 'chef-splunk::disabled'
-  Chef::Log.debug('Splunk is disabled on this node.')
-  return
+# Sets server type for configuring search server.
+node.default[:splunk][:type] = 'search'
+include_recipe 'chef-splunk::splunk_includes'
+
+# Creates .ui_login to bypass change password prompt
+login_ui = ::File.join(splunk_dir, 'etc', '.ui_login')
+cookbook_file login_ui do
+  source '.ui_login'
 end
 
-if node[:splunk][:is_server]
-  include_recipe "chef-splunk::#{node[:splunk][:type]}"
-else
-  include_recipe 'chef-splunk::client'
+# Tempalte to set ui-prefs
+template node[:splunk][:conf_files][:ui_prefs][:file] do
+  source   node[:splunk][:conf_files][:ui_prefs][:erb]
+  owner    node[:splunk][:conf_files][:perms][:owner]
+  group    node[:splunk][:conf_files][:perms][:group]
+  mode     node[:splunk][:conf_files][:perms][:mode]
+  notifies :restart, 'service[splunk]'
 end