+++ recipes/splunk_secrets.rb
@@ -1,9 +1,9 @@
 # encoding: utf-8
-#
 # Cookbook Name:: chef-splunk
-# Recipe:: default
-#
+# Recipe:: splunk_secrets
 # Author: Bernardo Macias <bmacias84@gmail.com>
+#
+# All rights reserved - Do Not Redistribute
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,16 +16,22 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
 
-if node[:splunk][:disabled]
-  include_recipe 'chef-splunk::disabled'
-  Chef::Log.debug('Splunk is disabled on this node.')
-  return
-end
+# load data bag for splunk
+splunk_secrets =[node[:splunk][:secret], node[:splunk][:passwd]]
 
-if node[:splunk][:is_server]
-  include_recipe "chef-splunk::#{node[:splunk][:type]}"
-else
-  include_recipe 'chef-splunk::client'
-end
+splunk_secrets.each { |secret|
+  secret = ChefVault::Item.load(
+    secret[:data_bag],
+    secret[:data_bag_item]
+  )
+  splunk_file = ::File.join(splunk_dir, 'etc')
+  splunk_file = ::File.join(splunk_file, 'auth') if secret['file-name'] == node[:splunk][:secret][:file]
+  splunk_file = ::File.join(splunk_file, secret['file-name'])
+  # creates files from databag
+  file splunk_file do
+    content secret['file-content']
+    mode   0400
+    action :create_if_missing
+  end
+}